Move witness logic to a mixin
